+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - run_program:  Fixed problem with mutable default arguments issue.
 
 ### Changed
+- main:  Added program lock functionality to program.
 - run_program:  Replaced crt_srv_inst call with create_instance call.
 - main:  Refactored "if" statements in function.
 - run_program:  Changed variable name to standard naming convention.